glm_ocr/requirements_glm.txt
transformers>=4.45.0
torch>=2.1.0
torchvision
timm
tiktoken
Pillow
sentencepiece
accelerate
einops
